moved enrichment of server features to server side
